+++ packages/components/src/Cards/ButtonCard/index.tsx
@@ -28,7 +28,6 @@ const StyledButtonCard = styled(motion.button)`
   position: relative;
   overflow: hidden;
   border-radius: 1rem;
-  width: var(--ha-device-button-card-width);
   display: flex;
   flex-direction: column;
   align-items: stretch;
@@ -38,6 +37,7 @@ const StyledButtonCard = styled(motion.button)`
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
   transition: var(--ha-transition-duration) var(--ha-easing);
   transition-property: background-color, box-shadow;
+  width: calc(100% - 2rem);
 
   &:disabled {
     cursor: not-allowed;
@@ -51,33 +51,44 @@ const StyledButtonCard = styled(motion.button)`
     background-color: var(--ha-S400);
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1);
   }
-  ${mq(
-    ["mobile", "tablet", "smallScreen"],
-    `
-    width: calc(100% - 2rem);
-  `,
-  )}
+  
 `;
 const StyledRipples = styled(Ripples)`
+  flex-shrink: 1;
   ${mq(
     ["mobile"],
     `
     width: 100%;
-    flex-shrink: 1;
   `,
   )}
   ${mq(
     ["tablet"],
     `
-    width: calc(50% - var(--gap) / 2);
-    flex-shrink: 1;
+    width: calc(50% - var(--gap, 0rem) / 2);
   `,
   )}
   ${mq(
     ["smallScreen"],
     `
-    width: calc((100% - 2 * var(--gap)) / 3);
-    flex-shrink: 1;
+    width: calc((100% - 2 * var(--gap, 0rem)) / 3);
+  `,
+  )}
+  ${mq(
+    ["mediumScreen"],
+    `
+    width: calc((100% - 3 * var(--gap, 0rem)) / 4);
+  `,
+  )}
+  ${mq(
+    ["desktop"],
+    `
+    width: calc((100% - 5 * var(--gap, 0rem)) / 6);
+  `,
+  )}
+  ${mq(
+    ["largeDesktop"],
+    `
+    width: calc((100% - 7 * var(--gap, 0rem)) / 8);
   `,
   )}
 `;