+++ packages/components/src/Cards/TimeCard/index.tsx
@@ -13,7 +13,7 @@ const Card = styled(motion.div)`
   position: relative;
   overflow: hidden;
   border-radius: 1rem;
-  width: var(--ha-device-time-card-width);
+  width: calc(100% - 2rem);
   display: flex;
   flex-direction: column;
   align-items: stretch;
@@ -22,6 +22,8 @@ const Card = styled(motion.div)`
   background-color: var(--ha-S300);
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
   transition: all 0.2s cubic-bezier(0.06, 0.67, 0.37, 0.99);
+  flex-shrink: 1;
+
   &:hover {
     background-color: var(--ha-S400);
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1);
@@ -30,14 +32,30 @@ const Card = styled(motion.div)`
     ["mobile"],
     `
     width: calc(100% - 2rem);
-    flex-shrink: 1;
   `,
   )}
   ${mq(
     ["tablet", "smallScreen"],
     `
-    width: calc((50% - var(--gap) / 2) - 2rem);
-    flex-shrink: 1;
+    width: calc((50% - var(--gap, 0rem) / 2) - 2rem);
+  `,
+  )}
+  ${mq(
+    ["desktop"],
+    `
+    width: calc(((100% - 2 * var(--gap, 0rem)) / 3) - 2rem);
+  `,
+  )}
+  ${mq(
+    ["desktop", 'mediumScreen'],
+    `
+    width: calc(((100% - 2 * var(--gap, 0rem)) / 3) - 2rem);
+  `,
+  )}
+  ${mq(
+    ["largeDesktop"],
+    `
+    width: calc(((100% - 3 * var(--gap, 0rem)) / 4) - 2rem);
   `,
   )}
 `;
@@ -164,13 +182,13 @@ function _TimeCard({
   const dateSensor = useEntity("sensor.date", {
     returnNullIfNotFound: true,
   });
-  const parts = convertTo12Hour(timeSensor?.state ?? "00:00");
   const [formatted, amOrPm] = useMemo(() => {
+    const parts = convertTo12Hour(timeSensor?.state ?? "00:00");
     const hour = parts.find((part) => part.type === "hour");
     const minute = parts.find((part) => part.type === "minute");
     const amOrPm = parts.find((part) => part.type === "dayPeriod");
     return [`${hour?.value}:${minute?.value}`, amOrPm?.value];
-  }, [parts]);
+  }, [timeSensor?.state]);
   if (!dateSensor || !timeSensor) {
     return <Warning />;
   }