+++ packages/components/src/Cards/TriggerCard/index.tsx
@@ -20,7 +20,7 @@ const StyledTriggerCard = styled(motion.button)`
   position: relative;
   overflow: hidden;
   border-radius: 1rem;
-  width: var(--ha-device-scene-card-width);
+  width: calc(100% - 2rem);
   aspect-ratio: 2/0.74;
   display: flex;
   flex-direction: column;
@@ -31,6 +31,7 @@ const StyledTriggerCard = styled(motion.button)`
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
   transition: var(--ha-transition-duration) var(--ha-easing);
   transition-property: box-shadow, background-color;
+  flex-shrink: 1;
   &:disabled {
     cursor: not-allowed;
     opacity: 0.8;
@@ -39,26 +40,30 @@ const StyledTriggerCard = styled(motion.button)`
     background-color: var(--ha-S400);
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1);
   }
-  ${mq(
-    ["mobile", "tablet", "smallScreen"],
-    `
-    width: calc(100% - 2rem);
-  `,
-  )}
 `;
 const StyledRipples = styled(Ripples)`
   ${mq(
     ["mobile"],
     `
     width: 100%;
-    flex-shrink: 1;
   `,
   )}
   ${mq(
     ["tablet", "smallScreen"],
     `
-    width: calc(50% - var(--gap) / 2);
-    flex-shrink: 1;
+    width: calc(50% - var(--gap, 0rem) / 2);
+  `,
+  )}
+  ${mq(
+    ["desktop", 'mediumScreen'],
+    `
+    width: calc(((100% - 2 * var(--gap, 0rem)) / 3));
+  `,
+  )}
+  ${mq(
+    ["largeDesktop"],
+    `
+    width: calc(((100% - 3 * var(--gap, 0rem)) / 4));
   `,
   )}
 `;