+++ packages/core/src/hooks/useEntity/index.ts
@@ -1,7 +1,6 @@
 import { useEffect, useMemo, useState, useCallback } from "react";
-import { isEqual, omit } from "lodash";
+import { isEmpty, omit } from "lodash";
 import TimeAgo from "javascript-time-ago";
-
 import type {
   HassEntityWithApi,
   HassEntityCustom,
@@ -9,10 +8,12 @@ import type {
   EntityName,
 } from "@typings";
 import type { HassEntity } from "home-assistant-js-websocket";
-import { useHass, useApi } from "@core";
+import { useApi, useHistory, useSubscribeEntity } from "@core";
 import { useDebouncedCallback } from "use-debounce";
 import { getCssColorValue } from "@utils/colors";
 import { computeDomain } from "@utils/computeDomain";
+import { diff } from "deep-object-diff";
+import type { HistoryOptions } from "../useHistory";
 
 // English.
 import en from "javascript-time-ago/locale/en.json";
@@ -35,11 +36,17 @@ interface UseEntityOptions {
   /** The amount of time to throttle updates in milliseconds */
   throttle?: number;
   returnNullIfNotFound?: boolean;
+  historyOptions?: HistoryOptions;
 }
 
 const DEFAULT_OPTIONS: UseEntityOptions = {
   throttle: 150,
   returnNullIfNotFound: false,
+  historyOptions: {
+    hoursToShow: 24,
+    significantChangesOnly: true,
+    minimalResponse: true,
+  },
 };
 
 type UseEntityReturnType<
@@ -53,16 +60,15 @@ export function useEntity<
   E extends EntityName,
   O extends UseEntityOptions = UseEntityOptions,
 >(entity: E, options: O = DEFAULT_OPTIONS as O): UseEntityReturnType<E, O> {
-  const { throttle, returnNullIfNotFound } = {
+  const { throttle, returnNullIfNotFound, historyOptions } = {
     ...DEFAULT_OPTIONS,
     ...options,
   };
-  const { getEntity } = useHass();
-  const timeSensor = getEntity("sensor.time", true);
-  const matchedEntity = getEntity(entity, returnNullIfNotFound);
+  const getEntity = useSubscribeEntity(entity);
+  const matchedEntity = getEntity(returnNullIfNotFound);
   const domain = computeDomain(entity) as ExtractDomain<E>;
   const api = useApi(domain, entity);
-
+  const history = useHistory(entity, historyOptions);
   const formatEntity = useCallback((entity: HassEntity): HassEntityCustom => {
     const now = new Date();
     const then = new Date(
@@ -103,20 +109,20 @@ export function useEntity<
 
   useEffect(() => {
     setEntity((entity) => (entity === null ? null : formatEntity(entity)));
-  }, [formatEntity, timeSensor]);
+  }, [formatEntity]);
 
   useEffect(() => {
-    const foundEntity = getEntity(entity, true);
-    if (
-      foundEntity &&
-      !isEqual(
-        omit(foundEntity, "custom", "last_changed", "last_updated"),
-        omit($entity, "custom", "last_changed", "last_updated"),
-      )
-    ) {
-      debounceUpdate(foundEntity);
+    const foundEntity = getEntity(true);
+    if (foundEntity && $entity) {
+      const diffed = diff(
+        omit(foundEntity, "custom", "last_changed", "last_updated", "context"),
+        omit($entity, "custom", "last_changed", "last_updated", "context"),
+      );
+      if (!isEmpty(diffed)) {
+        debounceUpdate(foundEntity);
+      }
     }
-  }, [$entity, matchedEntity, debounceUpdate, entity, getEntity]);
+  }, [$entity, debounceUpdate, getEntity]);
 
   return useMemo(() => {
     if ($entity === null) {
@@ -125,7 +131,8 @@ export function useEntity<
     }
     return {
       ...$entity,
+      history,
       api,
     } as unknown as UseEntityReturnType<E, O>;
-  }, [$entity, api]);
+  }, [$entity, history, api]);
 }